Moved google code to top of file
I enlisted the help of the Carly who thinks most likely there is a
javascript error in the success.php file before that code is executing
which means it wont execute. He suggested I move the code as close to
the top of the file as poss which I did then I ordered a test and ,made
a payment the real time GA tracked me 100% through the process including
the success.php. Note I've only added the GA code not any of the other
snippets if we get that working we can add that back then